requirements.txt
python-telegram-bot[asyncio]==22.1
redis
python-dotenv

